+++ clease/montecarlo/averager.py
@@ -13,7 +13,7 @@ class Averager:
 
         if abs(self._ref_value) < 1e-6:
             self._ref_value = 1.0
-        self._n_samples = 0.0
+        self._n_samples = 0
         self._mean = 0.0
 
     def __iadd__(self, value):
@@ -30,7 +30,7 @@ class Averager:
             self._n_samples += value._n_samples
             return self
 
-        self._n_samples += 1.0
+        self._n_samples += 1
         self._mean += value / self._ref_value
         return self