+++ tests/montecarlo/test_concentration_observer.py
@@ -1,67 +1,97 @@
-import unittest
+import pytest
 from clease.montecarlo.observers import ConcentrationObserver
-from clease.datastructures import SystemChange
+from clease.datastructures import SystemChange, MCStep, SystemChanges
 from ase.build import bulk
 
 
-class TestConcObsrever(unittest.TestCase):
-    def test_update(self):
-        atoms = bulk("Au") * (7, 7, 7)
-        obs = ConcentrationObserver(atoms, element="Au")
-        self.assertAlmostEqual(obs.current_conc, 1.0)
+@pytest.fixture
+def make_step():
+    def _make_step(**kwargs):
+        input = {
+            "step": 0,
+            "energy": -1.5,
+            "move_accepted": True,
+            "last_move": [],
+        }
+        input.update(**kwargs)
+        return MCStep(**input)
 
-        changes = [
-            SystemChange(0, "Au", "Cu"),
-            SystemChange(1, "Au", "Cu"),
-            SystemChange(2, "Au", "Cu"),
-        ]
-        obs(changes)
-        N = len(atoms)
-        self.assertAlmostEqual(obs.current_conc, 1.0 - 3.0 / N)
-
-        changes = [SystemChange(0, "Cu", "Au")]
-        obs(changes)
-        self.assertAlmostEqual(obs.current_conc, 1.0 - 2.0 / N)
-
-    def test_peak(self):
-        atoms = bulk("Au") * (7, 7, 7)
-        obs = ConcentrationObserver(atoms, element="Au")
-        self.assertAlmostEqual(obs.current_conc, 1.0)
-
-        changes = [
-            SystemChange(0, "Au", "Cu"),
-            SystemChange(1, "Au", "Cu"),
-            SystemChange(2, "Au", "Cu"),
-        ]
-        new_conc = obs(changes, peak=True)
-        self.assertAlmostEqual(new_conc, 1.0 - 3.0 / len(atoms))
-        self.assertAlmostEqual(obs.current_conc, 1.0)
-
-    def test_reset(self):
-        atoms = bulk("Au") * (7, 7, 7)
-        obs = ConcentrationObserver(atoms, element="Au")
-        self.assertAlmostEqual(obs.current_conc, 1.0)
-
-        changes = [
-            SystemChange(0, "Au", "Cu"),
-            SystemChange(1, "Au", "Cu"),
-            SystemChange(2, "Au", "Cu"),
-        ]
-        new_conc = obs(changes)
-        avg = obs.get_averages()
-        expect = (1.0 + 1.0 - 3.0 / len(atoms)) * 0.5
-        self.assertAlmostEqual(avg["conc_Au"], expect)
-
-        obs.reset()
-        avg = obs.get_averages()
-        self.assertAlmostEqual(avg["conc_Au"], 1.0 - 3.0 / len(atoms))
-
-    def test_none(self):
-        atoms = bulk("Au") * (7, 7, 7)
-        obs = ConcentrationObserver(atoms, element="Au")
-        conc = obs(None)
-        self.assertAlmostEqual(conc, 1.0)
+    return _make_step
 
 
-if __name__ == "__main__":
-    unittest.main()
+def test_update(make_step):
+    atoms = bulk("Au") * (7, 7, 7)
+    obs = ConcentrationObserver(atoms, element="Au")
+    assert obs.current_conc == pytest.approx(1.0)
+
+    changes = [
+        SystemChange(0, "Au", "Cu"),
+        SystemChange(1, "Au", "Cu"),
+        SystemChange(2, "Au", "Cu"),
+    ]
+    step = make_step(last_move=changes)
+    obs.observe_step(step)
+    N = len(atoms)
+    assert obs.current_conc == pytest.approx(1.0 - 3.0 / N)
+
+    changes = [SystemChange(0, "Cu", "Au")]
+    step = make_step(last_move=changes)
+    obs.observe_step(step)
+    assert obs.current_conc == pytest.approx(1.0 - 2.0 / N)
+
+    # Move was rejected, no change to the concentration
+    changes = [SystemChange(0, "Cu", "Au")]
+    step = make_step(last_move=changes, move_accepted=False)
+    obs.observe_step(step)
+    assert obs.current_conc == pytest.approx(1.0 - 2.0 / N)
+
+
+def test_peak(make_step):
+    atoms = bulk("Au") * (7, 7, 7)
+    obs = ConcentrationObserver(atoms, element="Au")
+    assert obs.current_conc == pytest.approx(1.0)
+
+    changes = [
+        SystemChange(0, "Au", "Cu"),
+        SystemChange(1, "Au", "Cu"),
+        SystemChange(2, "Au", "Cu"),
+    ]
+    step = make_step(last_move=changes)
+    new_conc = obs.observe_step(step, peak=True)
+    assert new_conc == pytest.approx(1.0 - 3.0 / len(atoms))
+    assert obs.current_conc == pytest.approx(1.0)
+
+
+def test_reset(make_step):
+    atoms = bulk("Au") * (7, 7, 7)
+    obs = ConcentrationObserver(atoms, element="Au")
+    assert obs.current_conc == pytest.approx(1.0)
+
+    changes = [
+        SystemChange(0, "Au", "Cu"),
+        SystemChange(1, "Au", "Cu"),
+        SystemChange(2, "Au", "Cu"),
+    ]
+    step = make_step(last_move=changes)
+    obs.observe_step(step)
+    avg = obs.get_averages()
+    expect = (1.0 + 1.0 - 3.0 / len(atoms)) / 2
+    assert avg["conc_Au"] == pytest.approx(expect)
+
+    obs.reset()
+    avg = obs.get_averages()
+    assert avg["conc_Au"] == pytest.approx(1.0 - 3.0 / len(atoms))
+
+
+def test_rejected(make_step):
+    atoms = bulk("Au") * (7, 7, 7)
+    obs = ConcentrationObserver(atoms, element="Au")
+    changes = [
+        SystemChange(0, "Au", "Cu"),
+        SystemChange(1, "Au", "Cu"),
+        SystemChange(2, "Au", "Cu"),
+    ]
+    step = make_step(move_accepted=False, last_move=changes)
+    conc = obs.observe_step(step)
+    # Move was rejected, so no change should be seen in the concentration
+    assert conc == pytest.approx(1.0)